Move packages into module
